refactor: move aggregate operators inside the api package
