+++ dsl/src/commonTest/kotlin/it/unibo/collektive/aggregate/SharingTest.kt
@@ -4,8 +4,8 @@ import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 import it.unibo.collektive.Collektive.Companion.aggregate
 import it.unibo.collektive.IntId
-import it.unibo.collektive.aggregate.ops.share
-import it.unibo.collektive.aggregate.ops.sharing
+import it.unibo.collektive.aggregate.api.operators.share
+import it.unibo.collektive.aggregate.api.operators.sharing
 import it.unibo.collektive.field.Field
 import it.unibo.collektive.field.max
 import it.unibo.collektive.field.min
